Add GetAll Method to Add

+++ Convertidor/Data/Entities/Adm/ADM_H_ORDEN_PAGO.cs
@@ -45,7 +45,7 @@
         public string? EXTRA13 { get; set; } = string.Empty;
         public string? EXTRA14 { get; set; } = string.Empty;
         public string? EXTRA15 { get; set; } = string.Empty;
-        public int? NUMERO_COMPROBANTE { get; set; }
+        public decimal? NUMERO_COMPROBANTE { get; set; }
         public DateTime? FECHA_COMPROBANTE { get; set; }
        
     }
